python-export/requirements.txt
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.7
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7